errores de compilacion solucionado

+++ src/main/java/com/dentalflow/pe/clinicalStaff/service/ClinicalStaffService.java
@@ -13,8 +13,8 @@ public interface ClinicalStaffService {
 
     @WebMethod
     ClinicalStaff createDentist(
-            Long userId,
-            Long specialtyId,
+            int userId,
+            int specialtyId,
             String licenseNumber,
             String firstName,
             String lastName,
