+++ src/main/java/com/dentalflow/pe/specialty/repository/ISpecialtyRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface ISpecialtyRepository
-        extends JpaRepository<Specialty, Long> {
+        extends JpaRepository<Specialty, Integer> {
 
     Optional<Specialty> findByName(String name);
 }
